hyperv/storvsc: tune ringbuffer, max_io and channel selection
max_io impact the storage performance for multiple threads.
ringbuffer is a factor to limit max_io.

Considering LUN and CPUID when selecting outgoing
channel. For RAID0 with many disks, LUN is the main
factor for channel selection. But if there is only
1 disk, CPUID is the main factor.

Submitted by: Hongjiang Zhang <honzhan microsoft com>